GUISignal

Description

GUISignal is an internal implementation module of HWebUI. HWebUI is providing FRP-based GUI functionality for Haskell by utilizing the Web-Browser. See module HWebUI for main documentation.

Documentation

data GUISignal

A GUI Signal is either an event coming from the GUI element upon user interaction or a command, send to the element

Constructors

GUIEvent GUIEvent

GUICommand GUICommand