GUIValue

Description

GUIValue is an internal implementation module of HWebUI. HWebUI is providing FRP-based GUI functionality for Haskell by utilizing the Web-Browser. See module HWebUI for main documentation.

Documentation

data GUIValue

A GUI value is the content of a GUI element, a String, a Bool a Number or similar. Compound types are realised as lists of values. This data type is used in GUIMessage to encode different values in a common data type.

Constructors

SVDouble Double

SVString String

SVList [GUIValue]

SVInt Int

SVBool Bool

SVEvent

SVNone