Widgets

Description

Widgets is an internal implementation module of HWebUI. HWebUI is providing FRP-based GUI functionality for Haskell by utilizing the Web-Browser. See module HWebUI for main documentation.

Documentation

wInitGUI

Arguments

:: Int
port used to communicate with Haskell server

-> HWebUIWidget
resulting Yesod widget

Yesod widget to initialize needed Javascript functionality in the HTML code of the GUI. Provides Dojokit inclusion and communication with Haskell Yesod server over websockets.

wButton

Arguments

:: String
Element Id

-> String
Label of the Button

-> HWebUIWidget
resulting Yesod Widget

Yesod widget for the Button GUI Element

wCheckBox

Arguments

:: String
Element Id

-> HWebUIWidget
resulting Yesod Widget

Yesod widget for the CheckBox GUI Element

wHtml

Arguments

:: String
Element Id

-> HWebUIWidget
resulting Yesod widget

Yesod widget for the PlainHtml GUI element (an element which is used for dynamic HTML output

wMultiSelect

Arguments

:: String
Element Id

-> Int
Width of widget

-> HWebUIWidget
resulting Yesod widget

Yesod widget for the MultiSelect GUI element

wNumberTextBox

Arguments

:: String
Element Id

-> HWebUIWidget
resulting Yesod widget

Yesod widget for the NumberTextBox GUI element

wRadioButton

Arguments

:: String
Element Id

-> String
Name

-> String
Value

-> Bool
Checked

-> HWebUIWidget
resulting Yesod widget

Yesod widget for the RadioButtion GUI element

wTextBox

Arguments

:: String
Element Id

-> HWebUIWidget
resulting Yesod widget

Yesod widget for the TextBox GUI element